+++ src/lib_gui/qt/utility/QtFlowLayout.cpp
@@ -41,7 +41,7 @@ int QtFlowLayout::verticalSpacing() const {
 }
 
 int QtFlowLayout::count() const {
-  return itemList.size();
+  return static_cast<int>(itemList.size());
 }
 
 QLayoutItem* QtFlowLayout::itemAt(int index) const {
